just use plain http for now
added try catch for sunrisesunset. also moved it to http vs https for
now.

seems i was doing that for weather stack anyway. makes the code clean
and consistent any way. if it becomes an issue, we could try https, if
fail, log and try http, if fail again, log an move on.